Move pass tester helper functions to single common place.
